pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dbx-ai-aviation"
version = "2.0.0"
description = "Production AI system for aviation safety analysis"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "DBX AI Team", email = "team@dbx-ai.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "psycopg2-binary>=2.9.0",
    "redis>=4.5.2",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "xgboost>=2.0.3",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0"
]
test = [
    "httpx>=0.24.0",
    "pytest-mock>=3.11.0"
]
monitoring = [
    "prometheus-client>=0.17.0",
    "structlog>=23.1.0"
]

[project.urls]
Homepage = "https://github.com/your-org/dbx-ai-aviation"
Documentation = "https://docs.dbx-ai.com"
Repository = "https://github.com/your-org/dbx-ai-aviation"
Issues = "https://github.com/your-org/dbx-ai-aviation/issues"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
